backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
aiohttp>=3.9.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-generativeai==0.3.1
groq==0.4.0
trafilatura==1.6.0 